Move endpoint registry API code to a separate package
