+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/endpoint/registry/impl/EndpointRegistryImpl.java
@@ -16,15 +16,15 @@
  * under the License.
  */
 
-package org.wso2.carbon.apimgt.impl;
+package org.wso2.carbon.apimgt.impl.endpoint.registry.impl;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.wso2.carbon.apimgt.api.APIManagementException;
-import org.wso2.carbon.apimgt.api.EndpointRegistry;
-import org.wso2.carbon.apimgt.api.model.EndpointRegistryEntry;
-import org.wso2.carbon.apimgt.api.model.EndpointRegistryInfo;
-import org.wso2.carbon.apimgt.impl.dao.ApiMgtDAO;
+import org.wso2.carbon.apimgt.impl.endpoint.registry.api.EndpointRegistry;
+import org.wso2.carbon.apimgt.impl.endpoint.registry.dao.EndpointRegistryDAO;
+import org.wso2.carbon.apimgt.impl.endpoint.registry.model.EndpointRegistryInfo;
+import org.wso2.carbon.apimgt.impl.endpoint.registry.model.EndpointRegistryEntry;
+import org.wso2.carbon.apimgt.impl.endpoint.registry.model.EndpointRegistryException;
 import org.wso2.carbon.apimgt.impl.internal.ServiceReferenceHolder;
 import org.wso2.carbon.apimgt.impl.utils.APIUtil;
 import org.wso2.carbon.user.api.UserStoreException;
@@ -38,10 +38,11 @@ import java.util.List;
 public class EndpointRegistryImpl implements EndpointRegistry {
 
     private static final Log log = LogFactory.getLog(EndpointRegistryImpl.class);
-    private ApiMgtDAO apiMgtDAO = ApiMgtDAO.getInstance();
+    private EndpointRegistryDAO registryDAO = EndpointRegistryDAO.getInstance();
     private String username;
 
     public EndpointRegistryImpl(String username) {
+
         this.username = username;
     }
 
@@ -50,24 +51,25 @@ public class EndpointRegistryImpl implements EndpointRegistry {
      *
      * @param endpointRegistry EndpointRegistryInfo
      * @return registryId UUID of the created Endpoint Registry ID
-     * @throws APIManagementException if failed to add EndpointRegistryInfo
+     * @throws EndpointRegistryException if failed to add EndpointRegistryInfo
      */
-    public String addEndpointRegistry(EndpointRegistryInfo endpointRegistry) throws APIManagementException {
+    public String addEndpointRegistry(EndpointRegistryInfo endpointRegistry) throws EndpointRegistryException {
+
         String tenantDomain = MultitenantUtils
                 .getTenantDomain(APIUtil.replaceEmailDomainBack(endpointRegistry.getOwner()));
         int tenantId;
         try {
             tenantId = ServiceReferenceHolder.getInstance().getRealmService().getTenantManager()
                     .getTenantId(tenantDomain);
-            if (apiMgtDAO.isEndpointRegistryNameExists(endpointRegistry.getName(), tenantId)) {
-                APIUtil.handleResourceAlreadyExistsException("Endpoint Registry with name '" + endpointRegistry
-                        .getName() + "' already exists");
+            if (registryDAO.isEndpointRegistryNameExists(endpointRegistry.getName(), tenantId)) {
+                EndpointRegistryUtil.handleResourceAlreadyExistsException("Endpoint Registry with name '"
+                        + endpointRegistry.getName() + "' already exists");
             }
-            return apiMgtDAO.addEndpointRegistry(endpointRegistry, tenantId);
+            return registryDAO.addEndpointRegistry(endpointRegistry, tenantId);
         } catch (UserStoreException e) {
             String msg = "Error while retrieving tenant information";
             log.error(msg, e);
-            throw new APIManagementException("Error in retrieving Tenant Information while adding endpoint" +
+            throw new EndpointRegistryException("Error in retrieving Tenant Information while adding endpoint" +
                     " registry :" + endpointRegistry.getName(), e);
         }
     }
@@ -78,19 +80,20 @@ public class EndpointRegistryImpl implements EndpointRegistry {
      * @param registryId   Registry Identifier
      * @param tenantDomain
      * @return An EndpointRegistryInfo object related to the given identifier or null
-     * @throws APIManagementException if failed to get details of an Endpoint Registry
+     * @throws EndpointRegistryException if failed to get details of an Endpoint Registry
      */
     public EndpointRegistryInfo getEndpointRegistryByUUID(String registryId, String tenantDomain)
-            throws APIManagementException {
+            throws EndpointRegistryException {
+
         int tenantId;
         try {
             tenantId = ServiceReferenceHolder.getInstance().getRealmService().getTenantManager()
                     .getTenantId(tenantDomain);
-            return apiMgtDAO.getEndpointRegistryByUUID(registryId, tenantId);
+            return registryDAO.getEndpointRegistryByUUID(registryId, tenantId);
         } catch (UserStoreException e) {
             String msg = "Error while retrieving tenant information";
             log.error(msg, e);
-            throw new APIManagementException("Error in retrieving Tenant Information while retrieving endpoint" +
+            throw new EndpointRegistryException("Error in retrieving Tenant Information while retrieving endpoint" +
                     " registry given by id: " + registryId, e);
         }
     }
@@ -99,10 +102,11 @@ public class EndpointRegistryImpl implements EndpointRegistry {
      * Deletes an Endpoint Registry
      *
      * @param registryUUID Registry Identifier(UUID)
-     * @throws APIManagementException if failed to delete the Endpoint Registry
+     * @throws EndpointRegistryException if failed to delete the Endpoint Registry
      */
-    public void deleteEndpointRegistry(String registryUUID) throws APIManagementException {
-        apiMgtDAO.deleteEndpointRegistry(registryUUID);
+    public void deleteEndpointRegistry(String registryUUID) throws EndpointRegistryException {
+
+        registryDAO.deleteEndpointRegistry(registryUUID);
     }
 
     /**
@@ -114,19 +118,20 @@ public class EndpointRegistryImpl implements EndpointRegistry {
      * @param offset       Offset
      * @param tenantDomain
      * @return A list of EndpointRegistryInfo objects
-     * @throws APIManagementException if failed to get details of an Endpoint Registries
+     * @throws EndpointRegistryException if failed to get details of an Endpoint Registries
      */
     public List<EndpointRegistryInfo> getEndpointRegistries(String sortBy, String sortOrder, int limit, int offset,
-                                                            String tenantDomain) throws APIManagementException {
+                                                            String tenantDomain) throws EndpointRegistryException {
+
         int tenantId;
         try {
             tenantId = ServiceReferenceHolder.getInstance().getRealmService().getTenantManager()
                     .getTenantId(tenantDomain);
-            return apiMgtDAO.getEndpointRegistries(sortBy, sortOrder, limit, offset, tenantId);
+            return registryDAO.getEndpointRegistries(sortBy, sortOrder, limit, offset, tenantId);
         } catch (UserStoreException e) {
             String msg = "Error while retrieving tenant information";
             log.error(msg, e);
-            throw new APIManagementException("Error in retrieving Tenant Information while retrieving details of " +
+            throw new EndpointRegistryException("Error in retrieving Tenant Information while retrieving details of " +
                     "endpoint registries", e);
         }
     }
@@ -135,11 +140,13 @@ public class EndpointRegistryImpl implements EndpointRegistry {
      * {@inheritDoc}
      */
     public EndpointRegistryEntry getEndpointRegistryEntryByUUID(String registryId, String registryEntryUuid)
-            throws APIManagementException {
-        return apiMgtDAO.getEndpointRegistryEntryByUUID(registryEntryUuid);
+            throws EndpointRegistryException {
+
+        return registryDAO.getEndpointRegistryEntryByUUID(registryEntryUuid);
     }
 
     /**
+     * /**
      * Returns all entries belong to a given endpoint registry
      *
      * @param sortBy          Name of the sorting field
@@ -154,15 +161,16 @@ public class EndpointRegistryImpl implements EndpointRegistry {
      * @param version         The version of registry entry
      * @param exactNameMatch  Whether to perform exact search on name
      * @return A list of EndpointRegistryEntry objects
-     * @throws APIManagementException if failed to get entries of an Endpoint Registry
+     * @throws EndpointRegistryException if failed to get entries of an Endpoint Registry
      */
     public List<EndpointRegistryEntry> getEndpointRegistryEntries(String sortBy, String sortOrder, int limit,
                                                                   int offset, String registryId, String serviceType,
                                                                   String definitionType, String entryName,
                                                                   String serviceCategory, String version,
                                                                   boolean exactNameMatch)
-            throws APIManagementException {
-        return apiMgtDAO.getEndpointRegistryEntries(sortBy, sortOrder, limit, offset, registryId, serviceType,
+            throws EndpointRegistryException {
+
+        return registryDAO.getEndpointRegistryEntries(sortBy, sortOrder, limit, offset, registryId, serviceType,
                 definitionType, entryName, serviceCategory, version, exactNameMatch);
     }
 
@@ -171,14 +179,15 @@ public class EndpointRegistryImpl implements EndpointRegistry {
      *
      * @param registryEntry EndpointRegistryEntry
      * @return entryID UUID of the created Registry Entry
-     * @throws APIManagementException if failed to add EndpointRegistryEntry
+     * @throws EndpointRegistryException if failed to add EndpointRegistryEntry
      */
-    public String addEndpointRegistryEntry(EndpointRegistryEntry registryEntry) throws APIManagementException {
-        if (apiMgtDAO.isRegistryEntryNameExists(registryEntry)) {
-            APIUtil.handleResourceAlreadyExistsException("Endpoint Registry Entry with name '"
-                    + registryEntry.getName() + "' and version '" + registryEntry.getVersion() + "' already exists");
+    public String addEndpointRegistryEntry(EndpointRegistryEntry registryEntry) throws EndpointRegistryException {
+
+        if (registryDAO.isRegistryEntryNameExists(registryEntry)) {
+            EndpointRegistryUtil.handleResourceAlreadyExistsException("Endpoint Registry Entry with name '"
+                    + registryEntry.getName() + "' already exists");
         }
-        return apiMgtDAO.addEndpointRegistryEntry(registryEntry, username);
+        return registryDAO.addEndpointRegistryEntry(registryEntry, username);
     }
 
     /**
@@ -186,33 +195,35 @@ public class EndpointRegistryImpl implements EndpointRegistry {
      *
      * @param entryName     original name of the registry entry
      * @param registryEntry EndpointRegistryEntry
-     * @throws APIManagementException if failed to update EndpointRegistryEntry
+     * @throws EndpointRegistryException if failed to update EndpointRegistryEntry
      */
     public void updateEndpointRegistryEntry(String entryName, EndpointRegistryEntry registryEntry)
-            throws APIManagementException {
+            throws EndpointRegistryException {
+
         if (!entryName.equals(registryEntry.getName()) &&
-                apiMgtDAO.isRegistryEntryNameExists(registryEntry)) {
-            APIUtil.handleResourceAlreadyExistsException("Endpoint Registry Entry with name '"
+                registryDAO.isRegistryEntryNameExists(registryEntry)) {
+            EndpointRegistryUtil.handleResourceAlreadyExistsException("Endpoint Registry Entry with name '"
                     + registryEntry.getName() + "' already exists");
         }
-        apiMgtDAO.updateEndpointRegistryEntry(registryEntry, username);
+        registryDAO.updateEndpointRegistryEntry(registryEntry, username);
     }
 
     /**
      * Deletes an Endpoint Registry Entry
      *
      * @param entryId Registry Entry Identifier(UUID)
-     * @throws APIManagementException if failed to delete the Endpoint Registry Entry
+     * @throws EndpointRegistryException if failed to delete the Endpoint Registry Entry
      */
-    public void deleteEndpointRegistryEntry(String entryId) throws APIManagementException {
-        apiMgtDAO.deleteEndpointRegistryEntry(entryId);
+    public void deleteEndpointRegistryEntry(String entryId) throws EndpointRegistryException {
+
+        registryDAO.deleteEndpointRegistryEntry(entryId);
     }
 
     /**
      * {@inheritDoc}
      */
     public void updateEndpointRegistry(String registryId, String registryName, EndpointRegistryInfo
-            endpointRegistryInfo) throws APIManagementException {
+            endpointRegistryInfo) throws EndpointRegistryException {
 
         String tenantDomain = MultitenantUtils
                 .getTenantDomain(APIUtil.replaceEmailDomainBack(endpointRegistryInfo.getOwner()));
@@ -222,15 +233,15 @@ public class EndpointRegistryImpl implements EndpointRegistry {
                     .getTenantId(tenantDomain);
             // if another registry with the updated name already exists, fail the operation.
             if (!registryName.equals(endpointRegistryInfo.getName()) &&
-                    apiMgtDAO.isEndpointRegistryNameExists(endpointRegistryInfo.getName(), tenantId)) {
-                APIUtil.handleResourceAlreadyExistsException("Endpoint Registry with name '" + endpointRegistryInfo
-                        .getName() + "' already exists");
+                    registryDAO.isEndpointRegistryNameExists(endpointRegistryInfo.getName(), tenantId)) {
+                EndpointRegistryUtil.handleResourceAlreadyExistsException("Endpoint Registry with name '"
+                        + endpointRegistryInfo.getName() + "' already exists");
             }
-            apiMgtDAO.updateEndpointRegistry(registryId, endpointRegistryInfo, username);
+            registryDAO.updateEndpointRegistry(registryId, endpointRegistryInfo, username);
         } catch (UserStoreException e) {
             String msg = "Error while retrieving tenant information";
             log.error(msg, e);
-            throw new APIManagementException("Error in retrieving Tenant Information while updating " +
+            throw new EndpointRegistryException("Error in retrieving Tenant Information while updating " +
                     "endpoint registry with id :" + registryId, e);
         }
 
@@ -242,15 +253,16 @@ public class EndpointRegistryImpl implements EndpointRegistry {
      * @param entryId       Registry Entry Identifier(UUID)
      * @param registryEntry EndpointRegistryEntry
      * @return entryID UUID of the created Registry Entry
-     * @throws APIManagementException if failed to delete the Endpoint Registry Entry
+     * @throws EndpointRegistryException if failed to delete the Endpoint Registry Entry
      */
     public String createNewEntryVersion(String entryId, EndpointRegistryEntry registryEntry)
-            throws APIManagementException {
-        if (apiMgtDAO.isRegistryEntryNameAndVersionExists(registryEntry)) {
-            APIUtil.handleResourceAlreadyExistsException("Endpoint Registry Entry with name '"
+            throws EndpointRegistryException {
+
+        if (registryDAO.isRegistryEntryNameAndVersionExists(registryEntry)) {
+            EndpointRegistryUtil.handleResourceAlreadyExistsException("Endpoint Registry Entry with name '"
                     + registryEntry.getName() + "' and version '" + registryEntry.getVersion() + "' already exists");
         }
-        return apiMgtDAO.addEndpointRegistryEntry(registryEntry, username);
+        return registryDAO.addEndpointRegistryEntry(registryEntry, username);
     }
 
 }
