+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/endpoint/registry/model/EndpointRegistryInfo.java
@@ -15,8 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-package org.wso2.carbon.apimgt.api.model;
+package org.wso2.carbon.apimgt.impl.endpoint.registry.model;
 
 /**
  * Endpoint Registry Info Object.
@@ -90,26 +89,33 @@ public class EndpointRegistryInfo {
     }
 
     public String getUpdatedBy() {
+
         return updatedBy;
     }
 
     public void setUpdatedBy(String updatedBy) {
+
         this.updatedBy = updatedBy;
     }
 
     public String getCreatedTime() {
+
         return createdTime;
     }
 
     public void setCreatedTime(String createdTime) {
+
         this.createdTime = createdTime;
     }
 
     public String getLastUpdatedTime() {
+
         return lastUpdatedTime;
     }
 
     public void setLastUpdatedTime(String lastUpdatedTime) {
+
         this.lastUpdatedTime = lastUpdatedTime;
     }
 }
+
