+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.endpoint.registry/src/gen/java/org/wso2/carbon/apimgt/rest/api/endpoint/registry/SettingsApiService.java
@@ -7,7 +7,7 @@ import org.apache.cxf.jaxrs.ext.MessageContext;
 import org.apache.cxf.jaxrs.ext.multipart.Attachment;
 import org.apache.cxf.jaxrs.ext.multipart.Multipart;
 
-import org.wso2.carbon.apimgt.api.APIManagementException;
+import org.wso2.carbon.apimgt.impl.endpoint.registry.model.EndpointRegistryException;
 
 
 import java.util.List;
@@ -19,6 +19,6 @@ import javax.ws.rs.core.SecurityContext;
 
 import org.wso2.carbon.apimgt.rest.api.endpoint.registry.RegistriesApi.*;
 
-@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJAXRSCXFCDIServerCodegen", date = "2020-05-26T12:22:45.951+05:30[Asia/Colombo]")public interface SettingsApiService {
-        public Response getSettings(MessageContext messageContext) throws APIManagementException;
+@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJAXRSCXFCDIServerCodegen", date = "2020-05-26T18:59:50.365+05:30[Asia/Colombo]")public interface SettingsApiService {
+        public Response getSettings(MessageContext messageContext) throws EndpointRegistryException;
 }
